Create pkg/logassembler package
The idea here is to define the maximum data for a message, which is
managed by the option `multilogger-max-size`.
The pkg/logassembler package represents a very basic logger message
assembler which will take care of concatenating every received log
message and pass it as a single log message to the configured logging
drivers.